# Use `StoreAligned` not `Store` in WidenAsciiToUtf16