## Description

This PR is to improve the in-loop `write` logic in `WidenAsciiToUtf16`, the major change is to replace `StoreAligned` with `Store` inside the loop to reduce the penalty caused by split loads.

We are open to adjusting the implementation style for either path. Perf number attached in the comments.