@Ruihan-Yin 2023-08-02
# Perf numbers

Base: main (531ad9596fef94194744ccfcac7159d1c1ae5e57)
Diff: main + changes on WidenAsciiToUtf16

## Avx512

summary:
better: 5, geomean: 1.099
worse: 1, geomean: 1.022
total diff: 6

| Slower                                                                | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Tests.Perf_Encoding.GetChars(size: 512, encName: "utf-8") |      1.02 |            85.25 |            87.11 |         |

| Faster                                                                 | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Tests.Perf_Encoding.GetChars(size: 1024, encName: "utf-8") |      1.18 |           176.81 |           150.47 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 1024, encName: "ascii") |      1.14 |           152.16 |           133.17 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 16, encName: "ascii")   |      1.09 |            17.97 |            16.54 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 512, encName: "ascii")  |      1.06 |            79.47 |            74.72 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 16, encName: "utf-8")   |      1.03 |            25.79 |            24.94 |         |


## AVX

summary:
better: 3, geomean: 1.049
total diff: 3

No Slower results for the provided threshold = 1% and noise filter = 0.5 ns.

| Faster                                                                 | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Tests.Perf_Encoding.GetChars(size: 512, encName: "ascii")  |      1.07 |            81.68 |            76.09 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 1024, encName: "utf-8") |      1.05 |           174.54 |           165.92 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 1024, encName: "ascii") |      1.02 |           148.66 |           145.48 |         |

## SSE
summary:
better: 5, geomean: 1.065
total diff: 5

No Slower results for the provided threshold = 1% and noise filter = 0.5 ns.

| Faster                                                                 | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Tests.Perf_Encoding.GetChars(size: 512, encName: "ascii")  |      1.11 |           108.07 |            97.21 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 16, encName: "ascii")   |      1.10 |            17.74 |            16.14 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 1024, encName: "ascii") |      1.05 |           183.11 |           173.62 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 1024, encName: "utf-8") |      1.04 |           219.66 |           211.76 |         |
| System.Text.Tests.Perf_Encoding.GetChars(size: 512, encName: "utf-8")  |      1.02 |           119.36 |           116.59 |         |