@Ruihan-Yin 2023-08-04
> How does this change affect `*-arm64` targets? The PR introduces unconditional `.ExtractMostSignificantBits()` which is expected to regress its performance.

Thanks for pointing out. That was a mistake, changed to `VectorContainsNonAsciiChar` to make sure Arm64 won't be affected by the used of `ExtractMSB`.