@xtqqczze 2023-08-09
Perhaps we shouldn't assume the argument `pUtf16Buffer` is char aligned, as `System.Text.Encoding.GetChars` is a public API.

I'm not sure how to handle this, but `System.SpanHelpers.IndexOfNullCharacter(System.Char*)` does the following check:

https://github.com/dotnet/runtime/blob/9a5fa7445c47e65e1cd454a5255b671ed81b3ae4/src/libraries/System.Private.CoreLib/src/System/SpanHelpers.Char.cs#L539-L542