@MichalPetryka 2023-08-10
> Perhaps we shouldn't assume the argument `pUtf16Buffer` is naturally aligned (i.e. 2 byte), as `System.Text.Encoding.GetChars` is a public API.

AFAIR the guideline is that if the platform the code is running on works with unaligned memory, dotnet APIs should work too.