@xtqqczze 2023-08-10
> AFAIR the guideline is that if the platform the code is running on works with unaligned memory, dotnet APIs should work too.

@MichalPetryka We have existing code that does not account for this, e.g. `WidenLatin1ToUtf16`, see #90319.