@anthonycanino 2023-08-10
> Perhaps we shouldn't assume the argument `pUtf16Buffer` is naturally aligned (i.e. 2 byte), as `System.Text.Encoding.GetChars` is a public API.
> 
> I'm not sure how to handle this, but `System.SpanHelpers.IndexOfNullCharacter(System.Char*)` does the following check:
> 
> https://github.com/dotnet/runtime/blob/9a5fa7445c47e65e1cd454a5255b671ed81b3ae4/src/libraries/System.Private.CoreLib/src/System/SpanHelpers.Char.cs#L539-L542

@tannergooding based on some of the conversation we had, I am not sure how to proceed.

Perhaps we want to implement a fallback method that does not pin, and does not explicitly use `StoreAligned` and instead uses `Store` but skips the initial alignment if the char pointer is not naturally aligned?