@xtqqczze 2023-08-13
> Perhaps we shouldn't assume the argument `pUtf16Buffer` is naturally aligned (i.e. 2 byte)

See also comments at #90319.