@anthonycanino 2023-10-31
> Just checking up on the status of this PR, @anthonycanino have you had the chance to address @tannergooding's feedback?

@tannergooding how do we feel about this change now that we are moving to `ISimdVector`? Is it better to address the alignment with that change in one PR?